backend/requirements.txt
# Web 框架
Flask==3.0.3

# MongoDB 对象文档映射器 (ODM)
mongoengine==0.28.2

# 处理跨域资源共享 (CORS)
Flask-Cors==4.0.1

# 用于密码哈希
bcrypt==4.1.3

# 用于生成和验证 JSON Web Tokens (JWT)
PyJWT==2.8.0

# 用于从 .env 文件加载环境变量，保证配置安全
python-dotenv==1.0.1

# --- 第三方服务 ---

# 用于发送短信 (SMS)
twilio

# 用于地理位置计算
geopy

# 新增依赖
dnspython==2.6.1
certifi==2024.6.2 